requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ultralytics==8.3.49
opencv-python==4.10.0.84
pillow==11.0.0
python-multipart==0.0.20
wandb==0.19.1
numpy==1.26.4
jinja2==3.1.4